@@ -12,18 +12,19 @@
     "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
-    "@figma/plugin-typings": "^1.78.1",
-    "@sveltejs/vite-plugin-svelte": "^2.4.6",
-    "@tsconfig/svelte": "^5.0.2",
-    "adorable-css": "1.5.6",
+    "@figma/plugin-typings": "^1.106.0",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tsconfig/svelte": "^5.0.4",
+    "adorable-css": "1.6.2",
     "jszip": "^3.10.1",
-    "prettier": "^3.1.1",
-    "svelte": "^4.2.1",
-    "svelte-check": "^3.5.2",
-    "tslib": "^2.6.2",
-    "tsup": "^7.2.0",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.10",
-    "vite-plugin-singlefile": "^0.13.5"
+    "prettier": "^3.4.2",
+    "prettier-plugin-tailwindcss": "^0.6.9",
+    "svelte": "^5.18.0",
+    "svelte-check": "^4.1.4",
+    "tslib": "^2.8.1",
+    "tsup": "^8.3.5",
+    "typescript": "^5.7.3",
+    "vite": "^6.0.7",
+    "vite-plugin-singlefile": "^2.1.0"
   }
 }
