@@ -1,27 +1,39 @@
 export const OPTIONS = {
-  "type": "inlineStyle"
+  type: "inlineStyle",
 }
 
-export const makeNumber = (num:number) => Number(num).toFixed(2).replace(/^0+|\.00$|0+$/g, "") || "0"
+export const makeNumber = (num: number) =>
+  Number(num)
+    .toFixed(2)
+    .replace(/^0+|\.00$|0+$/g, "") || "0"
 
-export const isNumber = (value:string|number):value is number => +value === +value
-export const isValid = (value:string|number) => value === 0 || !!value
-export const px = (value:string|number) => (isNumber(value) && value !== 0) ? Math.round(value) + "px" : String(value)
-export const percent = (value:string|number) => (isNumber(value) && value !== 0) ? makeNumber(value) + "%" : String(value)
+export const isNumber = (value: string | number): value is number => +value === +value
+export const isValid = (value: string | number) => value === 0 || !!value
+export const px = (value: string | number) => (isNumber(value) && value !== 0 ? Math.round(value) + "px" : String(value))
+export const percent = (value: string | number) => (isNumber(value) && value !== 0 ? makeNumber(value) + "%" : String(value))
 
-export const hex = (num:number) => Math.round(num * 255).toString(16).padStart(2, "0")
+export const hex = (num: number) =>
+  Math.round(num * 255)
+    .toString(16)
+    .padStart(2, "0")
 
+export const makeInt = (num: number) => makeNumber(Math.round(num))
 
-export const makeInt = (num:number) => makeNumber(Math.round(num))
-
-export const makeHexColor = (r:number, g:number, b:number, a:number = 1) => {
+export const makeHexColor = (r: number, g: number, b: number, a: number = 1) => {
   let hexColor = [r, g, b].map(hex)
-  if (a === 1 && hexColor.every(h => h[0] === h[1])) hexColor = hexColor.map(h => h[0])
+  if (a === 1 && hexColor.every((h) => h[0] === h[1])) hexColor = hexColor.map((h) => h[0])
   return hexColor.join("")
 }
 
 const makeAdorableStyleColor = ({r, g, b}, opacity = 1) => `#${makeHexColor(r, g, b)}${opacity === 1 ? "" : makeNumber(opacity)}`
-const makeTailwindStyleColor = ({r, g, b}, opacity = 1) => `#${makeHexColor(r, g, b, opacity)}${opacity === 1 ? "" : Math.round(+opacity * 255).toString(16).padStart(2, "0")}`
+const makeTailwindStyleColor = ({r, g, b}, opacity = 1) =>
+  `#${makeHexColor(r, g, b, opacity)}${
+    opacity === 1
+      ? ""
+      : Math.round(+opacity * 255)
+          .toString(16)
+          .padStart(2, "0")
+  }`
 
 export const makeColor = (color, opacity = 1) => {
   if (!color) return ""
@@ -32,8 +44,7 @@ export const makeColor = (color, opacity = 1) => {
   return makeTailwindStyleColor({r, g, b}, opacity)
 }
 
-
-export const makeGradientLinear = (paint:GradientPaint) => {
+export const makeGradientLinear = (paint: GradientPaint) => {
   // https://github.com/jiangyijie27/figma-copy-css-and-react-style/blob/master/code.ts
   const {gradientTransform, gradientStops} = paint as GradientPaint
   if (!gradientTransform || !gradientStops) {
@@ -60,7 +71,8 @@ export const makeGradientLinear = (paint:GradientPaint) => {
       m12: (gradientTransform[1][0] * gradientTransform[0][2] - gradientTransform[0][0] * gradientTransform[1][2]) * deltaVal,
     }
   }
-  const rotationTruthy = gradientTransformData.m00 * gradientTransformData.m11 - gradientTransformData.m01 * gradientTransformData.m10 > 0 ? 1 : -1
+  const rotationTruthy =
+    gradientTransformData.m00 * gradientTransformData.m11 - gradientTransformData.m01 * gradientTransformData.m10 > 0 ? 1 : -1
 
   const data = gradientTransformData
   const param = {x: 0, y: 1}
@@ -69,7 +81,9 @@ export const makeGradientLinear = (paint:GradientPaint) => {
     y: data.m10 * param.x + data.m11 * param.y,
   }
   const rad = makeNumber((Math.atan2(rotationData.y * rotationTruthy, rotationData.x * rotationTruthy) / Math.PI) * 180)
-  const gradientColors = gradientStops.map((gradient) => `${makeColor(gradient.color, gradient.color.a)} ${makeNumber(gradient.position * 100)}%`)
+  const gradientColors = gradientStops.map(
+    (gradient) => `${makeColor(gradient.color, gradient.color.a)} ${makeNumber(gradient.position * 100)}%`
+  )
 
   return `linear-gradient(${rad}deg,${gradientColors})`
 }
@@ -83,40 +97,56 @@ export const fourSideValues = (t, r, b, l) => {
 
 export const makeFourSideValues = (t, r, b, l) => fourSideValues(t, r, b, l).join("/")
 
-export const stripZero = (value:string) => value.startsWith("0.") ? value.slice(1) : value.startsWith("-0.") ? "-" + value.slice(2) : value
+export const stripZero = (value: string) =>
+  value.startsWith("0.") ? value.slice(1) : value.startsWith("-0.") ? "-" + value.slice(2) : value
 
-export const unitValue = ({value, unit}):string => {
-
-  console.warn(">>> 3")
+export const unitValue = ({value, unit}): string => {
   value = stripZero("" + makeNumber(value))
-
-  console.warn(">>> 4")
-
   switch (unit) {
-    case "PIXELS": {return px(value)}
-    case "PERCENT": {return percent(value)}
+    case "PIXELS": {
+      return px(value)
+    }
+    case "PERCENT": {
+      return percent(value)
+    }
   }
   return value
 }
 
 export const ab2str = (buf) => String.fromCharCode.apply(null, new Uint16Array(buf))
-export const capitalize = (str:string) => str.charAt(0).toUpperCase() + str.slice(1)
+export const capitalize = (str: string) => str.charAt(0).toUpperCase() + str.slice(1)
 
-export const nl2br = (str:string) => str.replace(/(\r\n|\n|\r|\u2028|\u2029)/g, "<br/>")
+export const nl2br = (str: string) => str.replace(/(\r\n|\n|\r|\u2028|\u2029)/g, "<br/>")
 
-export const indent = (code:string) => code ? "\n" + code.split("\n").map(line => "  " + line).join("\n") + "\n" : ""
+export const indent = (code: string) =>
+  code
+    ? "\n" +
+      code
+        .split("\n")
+        .map((line) => "  " + line)
+        .join("\n") +
+      "\n"
+    : ""
 
 export const traverse = (node, callback) => {
   callback(node)
   if (node.children && node.children.length) {
-    node.children.forEach(child => traverse(child, callback))
+    node.children.forEach((child) => traverse(child, callback))
   }
 }
 
-export const makeComponentName = (str:string) => capitalize(str.trim().replace(/[^_a-zA-Z0-9ㄱ-ㅎ가-힣]/g, "").replace(/\s*\/\s*/g, "_").replace(/-|\s+/g, "_").replace(/\s+/g, "_"))
+export const makeComponentName = (str: string) =>
+  capitalize(
+    str
+      .trim()
+      .replace(/[^_a-zA-Z0-9ㄱ-ㅎ가-힣]/g, "")
+      .replace(/\s*\/\s*/g, "_")
+      .replace(/-|\s+/g, "_")
+      .replace(/\s+/g, "_")
+  )
 
-
-export const makeFileName = (str:string) => str
-  .replace(/[\/\?<>\\:\*\|":]+/g, "_")
-  .replace(/\s+/g, "_")
-  .replace(/_+/g, "_")
+export const makeFileName = (str: string) =>
+  str
+    .replace(/[\/\?<>\\:\*\|":]+/g, "_")
+    .replace(/\s+/g, "_")
+    .replace(/_+/g, "_")
