@@ -1,9 +1,11 @@
-import "adorable-css"
-import './app.css'
-import App from './App.svelte'
+import {mount} from "svelte"
+// import "adorable-css"
+import "./adorable.css"
+import "./app.css"
+import App from "./App.svelte"
 
-const app = new App({
-  target: document.getElementById('app'),
+const app = mount(App, {
+  target: document.getElementById("app")!,
 })
 
 export default app
