@@ -1,61 +1,38 @@
-<!-- <Info_Top/> -->
-<div class="relative w(fill) vbox(center) layoutMode(VERTICAL)">
-  <div class="relative">
-    <svg width="80" height="80" viewBox="0 0 80 80" fill="none" xmlns="http://www.w3.org/2000/svg"
-         xmlns:xlink="http://www.w3.org/1999/xlink">
-      <mask id="mask0_527_95462" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="0" y="0" width="80" height="80">
-        <path
-          d="M79.9308 57.1877C79.5342 52.761 77.5275 48.5941 76.9411 44.1872C76.8764 43.7018 76.83 43.2164 76.7977 42.7295C76.6219 40.0964 77.041 37.4647 77.903 34.9713C78.8086 32.3495 79.2994 29.5485 79.2994 26.6388C79.298 11.9393 66.8177 0.0240193 51.3941 3.06277e-05C47.4412 -0.00561377 43.6795 0.76908 40.2708 2.17454C36.6455 3.66748 32.7292 4.40266 28.7819 4.28413C26.4616 4.21357 24.1427 4.31941 21.8379 4.675C14.1416 5.86174 6.50995 10.5028 4.35279 17.643C3.14483 21.6435 3.75233 25.905 3.64545 30.0649C3.4753 36.6773 1.48407 43.1105 0.508148 49.6594C-0.467777 56.2069 -0.306061 63.2991 3.41202 68.8843C6.10496 72.9285 10.453 75.7987 15.147 77.5103C19.8734 79.2347 25.0216 79.8344 30.0503 79.3363C33.7642 78.968 37.3711 77.7742 41.137 78.1058C50.4196 78.9242 57.8178 82.0865 66.9667 77.8109C74.9907 74.0616 80.6916 65.6726 79.9322 57.1877H79.9308Z"
-          fill="black"/>
-      </mask>
-      <g mask="url(#mask0_527_95462)">
-        <rect width="80" height="80" fill="url(#pattern0)"/>
-        <path
-          d="M79.7009 57.4236L79.6818 57.21C79.486 55.024 78.8921 52.8974 78.2671 50.7499C78.2283 50.6165 78.1893 50.483 78.1504 50.3494C77.564 48.3397 76.9711 46.3075 76.6933 44.2202C76.6278 43.7287 76.5809 43.2377 76.5482 42.7461M79.7009 57.4236L76.7977 42.7295M79.7009 57.4236C80.3355 65.7118 74.7407 73.9024 66.8609 77.5844C62.3374 79.6984 58.2476 79.9754 54.1235 79.6195C52.1005 79.445 50.0722 79.1184 47.9757 78.7809L47.8422 78.7594C45.701 78.4148 43.4898 78.0623 41.159 77.8568L41.137 78.1058L41.159 77.8568C39.2442 77.6882 37.3761 77.9079 35.5335 78.2119C34.9828 78.3028 34.436 78.4008 33.891 78.4985C32.6023 78.7295 31.3232 78.9588 30.0256 79.0875C25.0325 79.5821 19.9222 78.9864 15.2327 77.2755L15.147 77.5103L15.2327 77.2755C10.5725 75.5761 6.27523 72.7331 3.62012 68.7457C-0.0458805 63.2388 -0.21756 56.224 0.755417 49.6963L0.508148 49.6594M79.7009 57.4236L0.508148 49.6594M76.5482 42.7461L76.7977 42.7295M76.5482 42.7461C76.5482 42.7461 76.5483 42.7462 76.5483 42.7462L76.7977 42.7295M76.5482 42.7461C76.37 40.0756 76.7952 37.4104 77.6667 34.8896M76.7977 42.7295C76.6219 40.0964 77.041 37.4647 77.903 34.9713M77.6667 34.8896C77.6667 34.8896 77.6667 34.8897 77.6667 34.8897L77.903 34.9713M77.6667 34.8896C78.5634 32.2936 79.0494 29.5201 79.0494 26.6388C79.048 12.0884 66.6908 0.273822 51.3937 0.25003C47.4739 0.244433 43.7447 1.01261 40.3661 2.40566L40.366 2.4057C36.7089 3.91174 32.7576 4.65363 28.7744 4.53402L28.7743 4.53401C26.4666 4.46384 24.1632 4.5692 21.876 4.92208L21.876 4.92208C14.2312 6.10088 6.71061 10.7031 4.59211 17.7153M77.6667 34.8896L77.903 34.9713M77.903 34.9713L4.59211 17.7153M4.59211 17.7153C4.59211 17.7153 4.59211 17.7153 4.59212 17.7153L4.35279 17.643M4.59211 17.7153L21.8379 4.675C14.1416 5.86174 6.50995 10.5028 4.35279 17.643M4.59211 17.7153L4.35279 17.643M4.59211 17.7153C3.99832 19.6818 3.84864 21.7195 3.83621 23.7957C3.8308 24.6994 3.85135 25.6069 3.87199 26.5187C3.87507 26.6544 3.87814 26.7902 3.88113 26.9261C3.90419 27.9733 3.92226 29.0248 3.89537 30.0713C3.80974 33.3992 3.266 36.6774 2.62425 39.9343C2.44214 40.8585 2.25234 41.7802 2.06275 42.7008C1.58289 45.031 1.10438 47.3546 0.755418 49.6963L0.508148 49.6594M4.35279 17.643C3.48594 20.5138 3.554 23.519 3.62206 26.5247C3.64885 27.7074 3.67563 28.8902 3.64545 30.0649C3.53633 34.3056 2.67821 38.4726 1.81925 42.6436C1.33885 44.9764 0.858192 47.3105 0.508148 49.6594"
-          stroke="black" stroke-opacity="0.08" stroke-width="0.5"/>
-      </g>
-      <defs>
-        <pattern id="pattern0" patternContentUnits="objectBoundingBox" width="1" height="1">
-          <use xlink:href="#image0_527_95462" transform="scale(0.00446429)"/>
-        </pattern>
-        <image id="image0_527_95462" width="224" height="224" xlink:href=/>
-      </defs>
-    </svg>
+<!-- device=mobile, state=hovered -->
+<div class="w-[350px] h-[454px] flex flex-col gap-[20px] p-[20px] bg-[#fff] border-[2px_solid_#117ce9] rounded-[20px]" >
+
+  <!-- contents group1 -->
+  <div class="self-stretch flex-1 flex flex-col items-center justify-center overflow-hidden" >
+    <picture class="self-stretch w-fit flex-1" ></picture>
   </div>
-  <div class="relative w(fill) vbox(center) vgap(16) p(80/0/24) bg(#fff) r(24/24/0/0) layoutMode(VERTICAL)">
-    <div class="relative w(fill) vbox(left) vgap(8) p(0/32) layoutMode(VERTICAL)">
-      <div class="relative w(296) font(23/27) bold Noto-Sans-CJK-KR c(#141414) text(center)">
-        일이삼사오육칠팔구십일이삼사오육칠팔구십일이삼사오육칠팔구십
-      </div>
-      <div class="relative w(fill) h(16) hbox pack hgap(12) layoutMode(HORIZONTAL)">
-        <div class="relative hbox pack hgap(6) layoutMode(HORIZONTAL)">
-          <div class="relative">
-            <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path fill-rule="evenodd" clip-rule="evenodd"
-                    d="M5.99714 0.699951C5.49087 0.699951 5.08047 1.11036 5.08047 1.61662V2.61046C2.48064 3.04804 0.5 5.30969 0.5 8.03393V9.86727C0.5 10.3735 0.910406 10.7839 1.41667 10.7839H10.5833C11.0896 10.7839 11.5 10.3735 11.5 9.86727V8.03393C11.5 5.30769 9.51645 3.04472 6.9138 2.6095V1.61662C6.9138 1.11036 6.5034 0.699951 5.99714 0.699951ZM3.9375 5.97143C3.5578 5.97143 3.25 6.27924 3.25 6.65893V7.5756C3.25 7.9553 3.5578 8.2631 3.9375 8.2631C4.3172 8.2631 4.625 7.9553 4.625 7.5756V6.65893C4.625 6.27924 4.3172 5.97143 3.9375 5.97143ZM7.375 6.65893C7.375 6.27924 7.6828 5.97143 8.0625 5.97143C8.4422 5.97143 8.75 6.27924 8.75 6.65893V7.5756C8.75 7.9553 8.4422 8.2631 8.0625 8.2631C7.6828 8.2631 7.375 7.9553 7.375 7.5756V6.65893Z"
-                    fill="#1E96FF"/>
-            </svg>
-          </div>
-          <div class="relative font(13/16) Noto-Sans-CJK-KR c(#858585)">일이삼사오육칠팔구십일이</div>
+
+  <!-- Frame 1000011846 -->
+  <div class="self-stretch flex flex-col" >
+
+    <!-- Frame 1000011482 -->
+    <div class="w-fit flex flex-col gap-[8px]" >
+
+      <!-- Frame 1000010614 -->
+      <div class="h-[18px] shrink-0 flex flex-row gap-[6px]" >
+
+        <!-- tag_2.state -->
+        <div class="w-fit h-[18px] flex flex-row items-center justify-center p-[4px] bg-[#00d25a] rounded-[4px]" >
+          <div class="text-[12px] leading-[100%] font-['Pretendard'] font-semibold text-[#fff] text-center whitespace-nowrap" >Update</div>
         </div>
-        <div class="relative hbox pack hgap(6) layoutMode(HORIZONTAL)">
-          <div class="relative">
-            <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <circle cx="6" cy="3" r="2.5" fill="#66C959"/>
-              <path
-                d="M10.75 8.91431C10.75 10.3858 8.62335 10.2465 6 10.2465C3.37665 10.2465 1.25 10.3858 1.25 8.91431C1.25 7.44285 3.37665 6.25 6 6.25C8.62335 6.25 10.75 7.44285 10.75 8.91431Z"
-                fill="#66C959"/>
-            </svg>
-          </div>
-          <div class="relative font(13/16) Noto-Sans-CJK-KR c(#858585)">9,999명 플레이</div>
+
+        <!-- tag_4.info -->
+        <div class="w-fit h-[18px] flex flex-row items-center p-[4px] border-[1px_solid_#9e9e9e] rounded-[4px]" >
+          <div class="text-[12px] leading-[10px] font-['Pretendard'] font-semibold text-[#222] text-center whitespace-nowrap" >HPC</div>
         </div>
       </div>
-    </div>
-    <div
-      class="relative h(40) pack p(0/24) bg(linear-gradient(122.01deg,#31f7fd/23.96%,#95ccfc/50.52%,#f279f9/78.12%)) r(99) layoutMode(HORIZONTAL)">
-      <div class="relative font(14) bold Noto-Sans-CJK-KR c(#000)">입장하기</div>
+
+
+
+      <!-- Frame 2330 -->
+      <div class="self-stretch h-[136px] shrink-0 flex flex-col gap-[10px]" >
+        <H4 class="self-stretch text-[#222] max-w-[100%] overflow-hidden -webkit-box -webkit-box-orient-[vertical] line-clamp-1" >서비스 위한 아키텍처의 특징 게임 자체 스토리 대규모 서비스</H4>
+        <P2 class="self-stretch h-[72px] text-[#707070] max-w-[100%] overflow-hidden -webkit-box -webkit-box-orient-[vertical] line-clamp-1" >HPC(고성능 컴퓨팅)는 대규모 컴퓨팅 서비스로 외환 환율, 금융 파생 상품의 리스크를 분석하거나 기상청...</P2>
+      </div>
     </div>
   </div>
 </div>
-<!-- </Info_Top> -->
-
